fix bug, manager controller and golf controller

+++ back/src/controllers/manager.controller.js
@@ -4,18 +4,24 @@ const Golf = require('../models/golf.model');
 //CREER UN MANAGER
 exports.create = (req,res) => {
 
-	if(req.body.golf_id){
+	if("golf_id" in req.body){
 		Golf.findById(req.body.golf_id)
 			.then(golf =>{
-				if(golf.manager==false){
+				if(golf.manager_id==null){
 					const manager = new Manager({
 						name: req.body.name,
 						surname: req.body.surname,
 						email: req.body.email,
 						telephone: req.body.telephone,
-				        golf: true,
 				        golf_id: req.body.golf_id
-			        })					
+			        })	
+					manager.save().then(data=>{
+						res.send(data);
+					}).catch(err =>{
+						res.status(500).send({
+							message: err.message
+						})
+					})				        				
 				}
 				else{
 					res.send("Golf has already a manager")
@@ -33,7 +39,7 @@ exports.create = (req,res) => {
 			surname: req.body.surname,
 			email: req.body.email,
 			telephone: req.body.telephone,
-	        golf: false
+	        golf_id: null
 	        
 		})
 		manager.save().then(data=>{
@@ -48,12 +54,24 @@ exports.create = (req,res) => {
 
 // SUPPRIMER UN MANAGER
 exports.findOneAndRemove = (req, res) => {
-	Manager.findOneAndRemove({_id:req.params.id})
+	Manager.findById({_id:req.params.id})
 	.then(manager =>{
-		res.send("Remove Sucess")
-	})
+			if(manager.golf_id!=null){
+				res.status(500).send("Sorry but you can't remove a manager assign to a golf")
+			}else{
+				Manager.findOneAndRemove({_id:req.params.id})
+				.then(manager =>{
+					res.send("Remove Sucess")
+				})
+				.catch(err=>{
+					res.status(500).send({
+						message:err.message || "Error"
+					})
+				})
+			}
+		})
 	.catch(err=>{
-		res.status(500).send({
+	res.status(500).send({
 			message:err.message || "Error"
 		})
 	})
@@ -61,15 +79,20 @@ exports.findOneAndRemove = (req, res) => {
 
 // METTRE A JOUR UN MANAGER
 exports.findOneAndUpdate = (req,res) => {
-	Manager.findOneAndUpdate( {_id:req.params.id}, req.body)
-	.then(manager =>{
-		res.send("Update Succes");
-	})
-	.catch(err =>{
-		res.status(500).send({
-			message:err.message || "Some error occured when finding manager."
+	if("golf_id" in req.body){
+		res.status(500).send("Sorry but u can't change")
+	}
+	else{
+		Manager.findByIdAndUpdate( {_id:req.params.id}, req.body)
+			.then(manager =>{
+				res.send("Update Succes");
+			})
+			.catch(err =>{
+				res.status(500).send({
+					message:err.message || "Some error occured when finding manager."
+			})
 		})
-	})
+	}
 }
 
 // LIRE UN MANAGER
